Removed interval polling. Used ref for element animation. Moved opening state update to individual list components.
